+++ lib/screens/pageView/incomplete_tasks.dart
@@ -63,7 +63,7 @@ class _InCompleteTasksScreenState extends State<InCompleteTasksScreen> {
                         checkboxChange: (newValue) {
                           taskData.updateTask(taskData.tasks
                               .where((element) => !element.isDone!)
-                              .toList()[index] as int);
+                              .toList()[index] );
                         },
                         listTileDelete: () {
                           taskData.deleteTask(taskData.tasks
